configs/agent/requirements.txt
livekit-agents[openai,deepgram,cartesia,silero]==1.1.6
livekit-plugins-openai==1.1.6
livekit-plugins-deepgram==1.1.6
livekit-plugins-cartesia==1.1.6
livekit-plugins-silero==1.1.6
redis>=4.0.0
aioredis>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
pydantic>=2.0.0
loguru>=0.6.0
pytz>=2023.3
requests>=2.28.0